be more relaxed in what is accepted as a valid date format
This adheres to the principle of being strict in the data that is
generated but forgiving in accepting what is received.

This adds a few rudimentary tests.

The change here was originally made by @Klap-In for splitbrain/dokuwiki
in 4d1366e45cecd27c337b0a8d606cfcde58f9ac8c